Oxygen Concentration class
